requirements.txt
absl-py==0.4.0
astor==0.7.1
certifi==2018.8.13
chardet==3.0.4
dlib==19.15.0
gast==0.2.0
grpcio==1.14.1
h5py==2.8.0
idna==2.7
imageio==2.3.0
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
Markdown==2.6.11
numpy==1.14.5
opencv-python==3.4.2.17
Pillow==5.2.0
protobuf==3.6.1
PyYAML==3.13
requests==2.19.1
scikit-learn==0.19.2
scipy==1.1.0
six==1.11.0
tensorboard==1.10.0
tensorflow==1.10.0
termcolor==1.1.0
urllib3==1.23
Werkzeug==0.14.1
youtube-dl==2018.8.4
